requirements.txt
fastapi
uvicorn
chromadb
llama-cpp-python
duckduckgo-search
trafilatura
html2text
semantic-text-splitter
numpy
requests
